abstract out formatElapsedTime
